requirements.txt
django_pdb
django_extensions
python-social-auth
